Importing all my work from previous repo
Due to the repo owner not contributing a single LOC I am moving to my
own repo. So far I have constructed basic classes (some of which are not
being used and may be deleted later) and I have been working on the
input handling for rectangular selection of tiles on the board.