Added symbol drawing. Added skeleton for game logic.
Symbol drawing is done in Tile class according to its Symbol member
(which is assigned in test code in Board constructor. I should think
about how to assign symbols later though user won't be able to do that.
Also added a skeleton for game logic (GameLogic class) which takes board
as a member and checks it for Tatamibari rule compliance.